sharing services with extension
